+++ packages/contracts/contracts/interfaces/IJoinSplitVerifier.sol
@@ -1,9 +1,11 @@
 //SPDX-License-Identifier: GPL-3.0
 pragma solidity ^0.8.5;
 
+import {IBatchVerifier} from "./IBatchVerifier.sol";
+
 /// @title Verifier interface.
 /// @dev Interface of JoinSplit Verifier contract.
-interface IJoinSplitVerifier {
+interface IJoinSplitVerifier is IBatchVerifier {
     function verifyProof(
         uint256[2] memory a,
         uint256[2][2] memory b,