+++ packages/contracts/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- Add batch proof verification
 - Migrate for joinsplit
   - Compute operationDigest once for the entire operation
 - Rename all "flax" instances to "nocturne"